remove cyclic and rename files
