+++ src/renderer/components/docking/platform-dock-layout-sizing.tsx
