+++ src/renderer/components/docking/platform-dock-layout-update.tsx
@@ -1,6 +1,5 @@
 import LogError from '@shared/log-error.model';
 import {
-  TabInfo,
   isTab,
   SavedTabInfo,
   Layout,
@@ -8,37 +7,12 @@ import {
 } from '@shared/models/docking-framework.model';
 import DockLayout, { FloatPosition, TabData, PanelData } from 'rc-dock';
 import { TAB_TYPE_WEBVIEW } from '../web-view.component';
-import { TAB_GROUP, layoutDefaults, getFloatPosition } from './dock-layout-layout';
-import { loadTab } from './dock-layout-tab-storage';
+import { layoutDefaults, getFloatPosition } from './platform-dock-layout-sizing';
+import { loadTab } from './platform-dock-layout-tab-storage';
 import { TAB_TYPE_ERROR, ErrorTabData } from './error-tab.component';
-import PlatformPanel from './platform-panel.component';
-import PlatformTabTitle from './platform-tab-title.component';
 
 let previousTabId: string | undefined;
 let previousFloatPosition: FloatPosition = { left: 0, top: 0, width: 0, height: 0 };
-
-/**
- * Creates a tab ready to go into rc-dock from platform tab info
- *
- * @param tabInfo Data used to create the rc-dock tab
- * @returns Rc-dock tab created from `tabInfo`
- */
-export function createRCDockTabFromTabInfo(tabInfo: TabInfo) {
-  // Translate the data from the loaded tab to be in the form needed by rc-dock
-  return {
-    ...tabInfo,
-    title: (
-      <PlatformTabTitle
-        iconUrl={tabInfo.tabIconUrl}
-        text={tabInfo.tabTitle}
-        tooltip={tabInfo.tabTooltip}
-      />
-    ),
-    content: <PlatformPanel>{tabInfo.content}</PlatformPanel>,
-    group: TAB_GROUP,
-    closable: true,
-  };
-}
 
 /** Find the previous tab if one was previously added or find the first tab otherwise */
 export function findPreviousTab(dockLayout: DockLayout) {
